Add Support for upgrades (#117)
* Add Support for upgrades

When specifying a prefixed UUID Token on Acceptance Job, it fixes the path and docker images. This will give the possibility to upgrade instances smoothly without doing undeploy
Once the INSTANCE_TOKEN is specified. The naming of instance will use that token instead of using PLF Version which can be changed.

* Remaining side effect: The instance must configured with MySQL. This is not compatible with Postgres Upgrades (ToDo)

* Add Postgres Upgrade Phase

* Add Migration cleanup phase

* Update host namming