Upgrade Keycloak to 26.1 and Implement API-Based Permanent Admin User Creation (#257)
Keycloak now secures the bootstrap admin user (the user created by the docker run command), displaying a warning upon connection and treating it as a temporary user. As a result, the root username should no longer be created via the Docker CLI, as it was previously. This change updates the Keycloak version to 26.1 (the latest as of today) and introduces a new approach where the root user is created using its API services during the initial data setup.